[Spring 지하철 노선도 - 3단계] 릭(나일성) 미션 제출합니다. (#261)
* refactor(LineDaoTest): Assertj API를 활용해서 가독성 향상

* refactor(dao): RowMapper 로직 수정

* chore(DuplicateStationException): 컨벤션에 맞춰 공백 수정

* docks(README): 기능 구현 목록

* feat(schema.sql): 테이블 구조 변경

* feat(Section): 생성자

* feat(JdbcSectionDao): 구간 객체 저장 기능

* test(LineAcceptanceTest): 지하철 노선과 구간 생성 테스트

* feat(StationDao): id에 해당하는 역 조회

* feat(LineService): 노선과 구간을 저장하는 기능

* feat(LineController): 노선과 구간을 저장하는 기능

* feat(StationDao): 노선 아이디에 해당하는 모든 역 조회

* feat(LineService): id에 해당하는 노선 조회

* refactor(LineAcceptanceTest): id에 해당하는 노선 조회 기능 테스트 수정

* feat(LineService): 모든 노선 조회

* refactor(LineAcceptanceTest): 테스트 코드 리팩토링

* feat(SectionDao): 역이 구간에 등록되어 있는지 확인

* feat(StationService): 역 삭제 제약 조건

* refactor(LineRequest): 불필요한 생성자 삭제

* feat(SectionCreationRequest): 구간 생성 요청 DTO

* feat(Section): 생성자 및 toString, eq&hc 재정의

* feat(SectionDao): 노선이 일치하고 상행이나 하행역이 일치하는 구간 조회

* fix(SectionDao): 노선이 일치하고 하행이 일치하는 구간을 조회하지 못하는 문제 해결

* feat(SectionDao): 노선과 상행이 일치하는 구간 조회

* feat(SectionDao): 노선과 하행이 일치하는 구간 조회

* feat(SectionDao): 노선 삭제

* feat(SectionCreationRequest): Entity 변환 메서드

* feat(Section): getter, 값 비교 메서드

* feat(Section): 구간을 2개로 나누는 기능

* feat(SectionService): 새로운 구간을 저장

* feat(SectionController): 구간 등록

* test(SectionAcceptanceTest): 구간 등록 테스트

* feat(SectionDao): 노선과 역이 일치하는 구간 조회

* feat(SectionDao): 노선과 일치하는 모든 구간 조회

* feat(Section): 구간 2개를 하나로 합치는 기능

* feat(Section): 상행이나 하행역으로 포함하고 있는지 확인하는 기능

* feat(SectionService): 구간 제거

* feat(SectionController): 구간 제거

* fix(SectionService): 구간 사이에 새로운 구간을 추가하고 기존 구간을 다시 추가하는 문제 해결

* refactor(SectionService): Optional API를 활용

* feat(Section): Comparable 구현

* feat(LineService): 노선에 포함된 역을 상행 -> 하행 순으로 정렬하는 기능

* fix(Section): compareTo 메서드 구현 수정

* feat(LineService): 노선에 포함된 역을 상행 -> 하행 순으로 정렬하는 기능

* refactor(LineService): 노선에 포함된 역을 정렬하는 기능 리팩토링

* chore: 컨벤션에 맞춰 수정

* feat(Sections): Service에 있던 정렬 기능 할당

* refactor(SectionDao): 리스트 대신 일급 컬렉션을 반환하도록 변경

* refactor(LineService): SectionDao 인터페이스 변경에 대응

* refactor(Sections): SectionService의 비즈니스 로직을 Sections에 할당

* test(Sections): 삭제할 수 있는 구간 조회 기능 테스트

* refactor(Stations): Station 일급 컬랙션

* refactor(Section): 변수 추출

* refactor(SectionDao): 불필요한 메서드 삭제

* refactor(toString): StringBuffer -> + 연산으로 변경

* refactor(SectionController): 구간 생성 성공시 204를 내려주도록 변경

* refactor(SectionController): 구간 삭제 성공시 204를 내려주도록 변경

* refactor(LineController): 노선 정보 수정 성공시 204를 내려주도록 변경

* refactor(Distance): VO로 포장

* refactor(Section): 생성자 체이닝

* refactor(Line): 생성자 체이닝

* refactor(Station): 생성자 체이닝

* refactor(Distance): Section에 존재하던 Distance관련 메서드 이동

* refactor(Section): getter 대신 메시지를 받도록 변경

* refactor: 메서드 이름 변경

* feat(build.gradle): spring validation 의존성 추가

* feat(dto): 요청 dto Bean Validation 어노테이션 적용

* feat(ui): 매개 변수에 @Valid 어노테이션 추가

* feat(ExceptionAdvice): MethodArgumentNotValidException 핸들링 메서드

* feat(ExceptionAdvice): HttpMessageNotReadableException 핸들링 메서드

* refactor: IllegalArgumentException -> IllegalInputException 을 던지도록 변경

* refactor: 상행 종점 계산 로직 수정

* feat(Sections): 상행 종점 역을 찾는 로직에 검증 로직 추가

* feat(SectionDomain): 구간 도메인 객체 생성

* feat(JdbcSectionDomainDao): 구간 도메인 객체를 반환하는 Doa

* feat(SectionsDomain): 구간 도메인 일급 컬렉션 추가

* feat(JdbcSectionDomainDao): 노선에 해당하는 모든 구간 도메인 객체의 일급 컬렉션을 조회하는 기능

* feat(JdbcLineDomainDao): 도메인 객체 Dao 추가

* feat(domain): 필드를 id -> 도메인 객체로 변경

* refactor(JdbcSectionDao): SQL 상수 추출

* refactor(test): 테스트 코드 리팩토링

* refactor(main): 프로덕션 코드 리팩토링

* refactor(domain): 테스트에서만 쓰이는 생성자 삭제

* refactor(LineDaoTest): 픽스처를 활용해서 리팩토링

* refactor(SectionDaoTest): 픽스처를 활용해서 리팩토링

* refactor(StationDaoTest): 픽스처를 활용해서 리팩토링

* refactor(Sections): 픽스처를 활용해서 리팩토링

* refactor(SectionTest): 픽스처를 활용해서 리팩토링

* refactor(LineServiceTest): 픽스처를 활용해서 리팩토링

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.station.StationRequest;
 import wooteco.subway.dto.station.StationResponse;
+import wooteco.subway.exception.IllegalInputException;
 import wooteco.subway.exception.station.DuplicateStationException;
 
 @Service
@@ -15,9 +17,11 @@ import wooteco.subway.exception.station.DuplicateStationException;
 public class StationService {
 
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(final StationDao stationDao) {
+    public StationService(final StationDao stationDao, final SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse create(final StationRequest request) {
@@ -36,6 +40,9 @@ public class StationService {
     }
 
     public void delete(final Long id) {
+        if (sectionDao.existStation(id)) {
+            throw new IllegalInputException("역이 구간에 등록되어 있습니다.");
+        }
         stationDao.deleteById(id);
     }
 }
